reorganize velocity.py
